[Neo] #GIC_400: add some gic400 distributor registers, see more: https://developer.arm.com/documentation/ddi0471/b/programmers-model/distributor-register-summary

+++ Kernel/include/dlist.h
@@ -13,7 +13,7 @@ typedef struct ListNode {
 	struct ListNode *next;
 } ListNode;
 
-#define OffsetOf(type, member) (char *) (&(((type *) 0)->member))
+#define OffsetOf(type, member) ((int)((char *)&((type *)0)->member))
 #define ContainerOf(ptr, type, member) ((type *) ((char *) (ptr) - (char *) (&(((type *) 0)->member))))
 
 #endif /* __SYNESTIAOS_KERNEL_DLIST__ */
